requirements.txt
# --- Core web ---
Flask==3.1.0
Werkzeug==3.1.3
Flask-WTF==1.2.1
Flask-Limiter==3.12
flask-talisman==1.1.0
python-dotenv>=1.0.1

# --- PDF / imagens ---
PyPDF2<3
pikepdf>=9.0,<10
pypdfium2==4.30.0
Pillow==11.1.0
python-magic==0.4.27

# --- Tabelas / planilhas (lattice-first) ---
camelot-py==0.10.1
opencv-python-headless==4.8.1.78
ghostscript==0.8.1
pdfplumber==0.11.7
pandas==2.3.1
numpy==2.3.2
openpyxl==3.1.5
tabulate==0.9.0
pdfminer.six==20250506
pdf2docx==0.5.8